Add functions for each operation and result
I added a function for each of the 4 possible operations. Then I added
another function which calls the function corresponding to the operator
when the result button is clicked.

Then I set up some variables to store the user input as well and
selected all the available buttons to add eventlisteners to them.
When a number button is pressed, the value is displayed and as soon as
an operator button is clicked, the whole number is stored in a variable.