Create variable to store input and push to array
I created an additional variable to store the user input and as soon as
an operator button is pressed, the content of the variable is stored in
an array and the variable is cleared to take another user input.

Upon pressing the = button, the function to perform the calculation is
called and returns the result, which is displayed in the second line of
the calculator.