Add check for new operator before = and div by 0
I added some code so in case an operator button is pressed again before
running the calculation by pressing =, the calculation is executed and
all the variables are update just as if the = button had been pressed.
This way it is possible to chain several numbers an operators and after
every new operator the result is displayed

I also added a message to be displayed in case a division by 0 is tried.